refactor(trusty-common): single shared tmux library; route trusty-mpm + trusty-agents through it (closes #3004, refs #2398 #2399)
Two independent tmux implementations existed in the workspace: trusty-mpm's
core::tmux (the #2398/#2399 choke point that correctly applies
history-limit/mouse BEFORE new-session) and trusty-agents' orchestrator.rs +
debugger/tmux.rs (an independent duplicate with no scrollback handling at
all). A scrollback bug report against trusty-agents surfaced the duplication
as the real root cause: #2399 could only fix the crate it lived in.

Extracts the pure tmux command/argv layer (TmuxTarget, TmuxCommand,
tmux_argv, scrollback_option_commands, the new managed_session_commands
ordering-guarantee recipe, DEFAULT_TMUX_HISTORY_LIMIT=100000/DEFAULT_TMUX_MOUSE)
into trusty-common::tmux as the single source of truth. trusty-mpm's
core::tmux becomes a thin re-export/adapter that keeps everything genuinely
mpm-specific (binary resolution, TCC-disclaimed process spawning, config-file
resolution). trusty-agents' orchestrator + debugger route their
create_session through the same shared recipe, so both now apply the
scrollback/mouse options before pane creation.
