+++ crates/trusty-common/CHANGELOG.md
@@ -9,6 +9,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- `tmux` module: the shared, pure tmux command-construction layer (`TmuxTarget`, `TmuxCommand`, `tmux_argv`, `scrollback_option_commands`, `managed_session_commands`, `DEFAULT_TMUX_HISTORY_LIMIT`/`DEFAULT_TMUX_MOUSE`) — single workspace-wide source of truth for the history-limit/mouse-before-`new-session` ordering guarantee, so trusty-mpm and trusty-agents no longer maintain two independent tmux implementations (closes [#3004](https://github.com/bobmatnyc/trusty-tools/issues/3004), refs #2398, #2399)
 - `sld` module (behind the new lightweight `sld` feature — `regex` + `serde_yaml` + `thiserror`, deliberately NOT the heavy `intent-source`/`tickets` stack): the language-agnostic **Spec-Linked Documentation (DOC-38)** reference grammar — the `SPEC-{SUBSYSTEM}-{NN}~{rev}` id grammar + §2.2 reference regex, the per-extension comment-syntax table, fenced-code-aware inline `# Spec References` block parsing, `spec_refs:` YAML-frontmatter parsing, and `{#SPEC-…}` heading-anchor scanning/resolution. Consolidates `revision_of`/`base_id` here as the single source both this grammar and `intent_source::spec_resolve` share (the `intent-source` feature now enables `sld`), so the new `trusty-sld-lint` gate and the ISR parse ONE grammar (DOC-38 §10 F1) ([#2854](https://github.com/bobmatnyc/trusty-tools/issues/2854))
 - `search_readiness::log_readiness`: emits the session-facing readiness log line for an ALREADY-probed `IndexReadiness`. `log_index_readiness` now delegates to it (identical behaviour), so a caller that must both log the readiness AND forward it elsewhere — e.g. `trusty-code` publishing it as a daemon API event for its UI — can probe ONCE rather than paying for a second HTTP probe that might disagree about a lane that flipped in between ([#2784](https://github.com/bobmatnyc/trusty-tools/issues/2784))
 - `search_readiness` module (behind the `search-index` feature): a shared trusty-search index readiness probe so a daily-driver session can *surface* whether a project's index is ready per lane (lexical/semantic/graph) instead of silently querying a not-yet-warm index — `parse_readiness` (pure status→snapshot map), `probe_index_readiness` (fail-open `GET /indexes/{id}/status`), and `log_index_readiness` (one stderr line) ([#2784](https://github.com/bobmatnyc/trusty-tools/issues/2784))
