+++ crates/trusty-mpm/CHANGELOG.md
@@ -17,6 +17,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - `write_project_hooks` (the daemon's managed-launch project-tier `.claude/settings.json` writer) now writes via `trusty_common::claude_config::write_json_atomic` (temp-file + rename, with a `.bak` backup of the prior content) instead of a plain `fs::write`, matching the atomic-write convention already used by the sibling `tm hooks clean` mutation path. Since #2969 this function writes the full lifecycle triad on every managed-session launch, so a crash or `^C` mid-write previously risked leaving a truncated `settings.json`; the write mechanism now can no longer tear (closes #2972)
+- `core::tmux`'s pure argv-construction layer (`TmuxTarget`, `TmuxCommand`, `tmux_argv`, `scrollback_option_commands`, `DEFAULT_TMUX_HISTORY_LIMIT`/`DEFAULT_TMUX_MOUSE`) moved to `trusty-common::tmux` as the workspace-wide single source of truth (issue #3004) — a scrollback bug report against trusty-agents' independent tmux implementation (which never got the #2398/#2399 history-limit-before-`new-session` fix) surfaced the real root cause: two crates were maintaining separate, drifting tmux command layers. `core::tmux` is now a thin re-export/adapter; `create_managed_session`'s behavior, config resolution, and TCC-disclaimed spawning are unchanged (closes #3004, refs #2398, #2399)
 - The bare `tm` guided default (no subcommand) no longer silently auto-starts or reconnects to a DIFFERENT daemon when an explicitly-supplied `--url` / `TRUSTY_MPM_URL` is unreachable — it now probes the explicit URL up front and exits 75 with a message naming the failed URL, instead of falling through to `ensure_daemon_started`'s implicit lock-file/default re-resolution and silently operating against whatever real daemon happens to be running. Implicit resolution (no `--url`/`TRUSTY_MPM_URL` set) and `tm hook`'s fail-open contract are both unaffected. `resolve_daemon_url_probing` (library-level) and the new `resolve_daemon_url_for_cli` helper now return a `DaemonUrlError::Unreachable` instead of falling back for an explicit, unreachable URL; the exit code (75) is a shared constant also used by the pre-existing `tm session prune-idle` SM-unavailable exit code (closes #1737)
 - `tm doctor`'s `output_style` check validated resolution only (configured id maps to a non-empty file) and missed content drift — a deployed style file could carry stale text (e.g. an already-corrected PM identity string) through an install + daemon restart while `tm doctor` still reported `Ok`. Adds a new `output_style_staleness` check that byte-compares each deployed `~/.claude/output-styles/*.md` file against the bundled catalog and `Warn`s on drift (pointing at `tm install`), plus orphan-file detection for foreign/dormant files in the same directory (e.g. a pre-rebrand `claude-mpm.md`) that match no bundled style and are not the currently configured id — flagged, never auto-deleted, bringing the doctor suite to 18 checks (closes #2333)
 - `tm install` no longer discovers and mutates every `.claude/settings.json` / `settings.local.json` reachable under `$HOME` — tm hooks are now written ONLY into the tm-owned managed `CLAUDE_CONFIG_DIR`, never into project directories. Adds `tm hooks clean [--path <dir>] [--force]` to remove tm hook entries a pre-fix install already wrote into project settings files (dry-run by default, backs up before an atomic temp-file+rename write, never a truncate-in-place — a kill mid-write can no longer corrupt `settings.json`), and two new `tm doctor` checks — `hooks_contamination` (tm hooks found in a known project's settings, suggests `tm hooks clean`) and `hooks_foreign_conflict` (foreign/claude-mpm hooks that would fire inside a tm session, informational only) — bringing the doctor suite to 17 checks. The hash-suffixed-binary hook matcher now additionally requires a `deps/` path component before `--force` will delete a match, narrowing the risk of misclassifying a coincidentally-named foreign binary (closes #2940) ([#2945](https://github.com/bobmatnyc/trusty-tools/pull/2945)) ([`3940d7f`](https://github.com/bobmatnyc/trusty-tools/commit/3940d7f90d25fd57ef02ea47bb09a20cca06d34e))
